remove coupling between pulsar and warp10
